requirements.txt
Flask
Flask-SocketIO
pyserial